requirements.txt
pwntools
python-dotenv
